Refactor storage tests using compaction
